rendering node directions properly
FlowFieldPathFinder now calculates the direction of each node with the function calculate flow field. It is having problems comparing numbers to infinity so currently when we need to use infinity we are using a very high number. Units are also moving to the center point of the flow field but then over shoot where they are supposed to go.