Almost Finished Player Movement
Changed Kinematic steering so that it is based off of the desired number of frames that you want it to take for the player to reach the next node. I was having a problem when I was using the already built in steering component velocity so I essential Divide the distance between the player and their destination with the desired number of frames, and increment their position based on this division. I want to change it so that it is velocity based