CollisionComponent
Added Collision Component. Similar to physics, steering, and position. Currently an if statement in its update is not working properly. It compares two IDs and when they are different it acts as if they are the same still. Removed Pathfinding inputs. Added a Unit type to compare what kind of unit is colliding with the other. Added a HandleComponentMessage that checks to see what is colliding with what.