Added SpawnCoinMessage
Collision still doesn't work. Always passes through the first unit then deletes it and goes through until it deletes the one that you are on top of. Thinks that the one you are colliding with is the first one. Spawn coin always make sure there are X amount of Coins on the screen, updates every frame.